requirements.txt
dash
gunicorn
psycopg2-binary
redis
numpy
faker